include all album types for artists
