+++ packages/server/src/spotify/api/helpers/getArtistAlbums.js
@@ -1,0 +1,10 @@
+const uniqueAlbums = require('../../utils/uniqueAlbums');
+
+module.exports = async function getArtistAlbums(spotifyApi, artistId) {
+  const data = await spotifyApi.getArtistAlbums(artistId);
+
+  return {
+    artistId,
+    albums: uniqueAlbums(data.body.items),
+  };
+};
